perf: make the eddblink import use an intermediate table
the eddblink import can be quite huge and directly writing every stationitem into the table has a massive overhead. this is an experiment at offsetting that by instead writing the items into a temporary table and then asking the database to transfer them into the original table; this way the database's own planner/optimizer has all the information it could possibly want in order to reason about the change.

this is a work-in-progress, tho: the transfer itself appears to be quite punishing, probably because its trying to do everything in a single thread/operation, so we may need additional work to split the station item table into sub-tables somehow.